+++ HISTORY.rst
@@ -34,6 +34,7 @@ TODO
 * Fetch variants using positions
 * Import from plink using *genomic coordinates*
 * Import *50K*, *600K* and *WGS isheep* datasets (`#47 <https://github.com/cnr-ibba/SMARTER-database/issues/47>`__)
+* Fix issue in ``src.features.plinkio.plink_binary_exists``
 
 0.4.5 (2022-06-14)
 ------------------